requirements.txt
gradio
timm
torch
torchvision
pillow
